Update prompt docs to describe 4th option (no value)
Supplying no value for the `prompt` arugment allows for the default
google prompt behavior. (Per the docs)[https://developers.google.com/identity/openid-connect/openid-connect#prompt])
"If no value is specified and the user has not previously authorized access, then the user is shown a consent screen.".
But if the user is only logged in to one google account and has already
consented, neither screen is shown the user is authenticated directly.
I've tested the behavior and it works as described.

I wasn't sure how best to phrase this new option in the doc. Since
`none` is a valid option, I thought it'd be a little unclear to
put both `none` and `None` in the docs, with the latter being the
python None type (no quotes). I make it clear, I added quotes to
the existing options and no quotes for python `None` type. You can
also achieve this new option with empty quotes `""`, so that could
be another option if the current update to the docs is not clear.